Split refresh logic into separate files
