+++ next-frontend/src/lib/wca/oauth/config.ts
@@ -1,0 +1,3 @@
+export const WCA_OIDC_ISSUER = process.env.OIDC_ISSUER!;
+export const WCA_OIDC_CLIENT_ID = process.env.OIDC_CLIENT_ID!;
+export const WCA_OIDC_CLIENT_SECRET = process.env.OIDC_CLIENT_SECRET!;
